load_hit/stamp returns offset as separate value
Instead of adding the file offset to the metadata, `find_hit` and
`load_stamp` now return the the file offset as a third tuple field.
This lets the caller decide if/when to store the file offset with the
metadata, which can help with uniqueness checks as in `find_hits`.