Major iteration update
The `CapnpReader` has several new constructors that allow for very
flexible iteration through a Capnp file:

- `CapnpReader(fname::AbstractString)`

  Iterating over an object created by calling this basic constructor
  will produce a sequence of `Tuple{Vector{UInt64}, Int64}`.  These
  Tuples can be passed to the Hit or Stamp constructor, as appropriate,
  to create a Hit or Stamp object.  Because that is so common, the
  constructor can be called with a Type parameter (see next
  constructor).

- `CapnpReader(::Type{T}, fname::AbstractString)`

  Each iteration of the object returned by calling this constructor will
  load an object of type `T` from the file and return it.  For this to
  work, type `T` must have a constructor that accepts a
  `Tuple{Vector{UInt64}, Int64}` and the file must contain a sequence of
  type `T`.  For `seticore` Hits and Stamps files, the `Hit` and `Stamp`
  types can be used for `T`.

- `CapnpReader(f::Function ::Type{T}, fname::AbstractString)`

  This is the most generic `CapnpReader` constructor.  Iterating over
  the returned object will call `f(T, t::Tuple{Vector{UInt64}, Int64})`
  and return the result.  In this context, `f` is a factory function
  that will load on object of type `T` from the file (actually the
  `Vector` and `Int64` of the tuple) and then return the object or some
  variation of it (e..g the object and the file offset from which it was
  read).  Factory methods can also call other constructor methods of
  type T, such as to pass keyword arguments.

The CapnpReader type is now parameterized with a type parameter `T` and
a factory function parameter `F`: `CapnpReader{T,F}`. The type parameter
indicates the type of object that will be constructed from the Capnp
file on each iteration and the factory function will be used to perform
the construction.  The factory function need not return (just) an
instance of the type, but that is the most basic/common case.  Several
factory functions are provided, but the user can define their own
factory functions if desired.  A factory function must have a method
with a signature that matches this:

    my_factory_function(::Type{T}, t::Tuple{Vector{UInt64}, Int64})

The `T` type may be an explicit type or a parameterized type.  If the
return type of the factory function is known, the user may provide a
method for `Base.IteratorEltype` that takes a
`::Type{CapnpReader{T,my_factory_function}}` type and returns
`Base.Eltype()` as well as a `Base.eltype` method that returns the type
returned by the factory function.  See the existing factory code in
`src/capnp.jl`.

The following factory methods are provided:

- `default_factory(::Type{Tuple}, t:Tuple{Vector{UInt64}, Int64})`

  Low level factory that simply returns `t`.

- `default_factory(::Type{T}, t:Tuple{Vector{UInt64}, Int64}) where T`

  Default factory that returns `T(t)`.

- `nodata_factory(::Type{T}, t:Tuple{Vector{UInt64}, Int64}) where T`

  Like `default_factory` but passes `withdata=false` to omit data.  The
  `withdata` keyword is supported by the `Filterbank`, `Hit` and `Stamp`
  constructors.  Returns an instance of `T`.

- `offset_factory(::Type{T}, t:Tuple{Vector{UInt64}, Int64}) where T`

  Like `nodata_factory` but returns a `Tuple{T,Int64}` where the `Int64`
  is the zero based byte offset of the start of the frame from which the
  object was read.

- `index_factory(::Type{T}, t:Tuple{Vector{UInt64}, Int64}) where T`
  Like `offset_factory` but the `Int64` is the one based word (`UInt64`)
  offset of the start of the frame from which the object was read.